requirements.txt
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
scikit-learn==1.3.0
tensorflow==2.13.0
matplotlib==3.7.2
seaborn==0.12.2
pyyaml==6.0
joblib==1.3.1
tqdm==4.65.0
wandb==0.15.5
pytest==7.4.0
jinja2==3.1.2